src/test/java/uj/jwzp/w2/e3/SellingServiceTest.java
package uj.jwzp.w2.e3;

import org.junit.Assert;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.junit.MockitoJUnit;
import org.mockito.junit.MockitoRule;
import org.powermock.api.mockito.PowerMockito;
import org.powermock.core.classloader.annotations.PrepareForTest;
import org.powermock.modules.junit4.PowerMockRunner;
import uj.jwzp.w2.e3.external.DiscountsConfig;
import uj.jwzp.w2.e3.external.PersistenceLayer;

import java.math.BigDecimal;

@RunWith(PowerMockRunner.class)
public class SellingServiceTest {

    @Mock
    private PersistenceLayer persistenceLayer;

    @Rule
    public MockitoRule mockitoRule = MockitoJUnit.rule();

//    @Test
//    public void notSell() {
//        //given
//        SellingService uut = new SellingService(persistenceLayer);
//        Mockito.when(persistenceLayer.saveCustomer(Mockito.any())).thenReturn(Boolean.TRUE);
//        Item i = new Item("i", new BigDecimal(3));
//        Customer c = new Customer(1, "DasCustomer", "Kraków, Łojasiewicza");
//
//        //when
//        boolean sold = uut.sell(i, 7, c);
//
//        //then
//        Assert.assertFalse(sold);
//        Assert.assertEquals(BigDecimal.valueOf(10), uut.moneyService.getMoney(c));
//    }
//
//    @Test
//    public void sell() {
//        //given
//        SellingService uut = new SellingService(persistenceLayer);
//        Mockito.when(persistenceLayer.saveCustomer(Mockito.any())).thenReturn(Boolean.TRUE);
//        Item i = new Item("i", new BigDecimal(3));
//        Customer c = new Customer(1, "DasCustomer", "Kraków, Łojasiewicza");
//
//        //when
//        boolean sold = uut.sell(i, 1, c);
//
//        //then
//        Assert.assertFalse(sold);
//        Assert.assertEquals(BigDecimal.valueOf(7), uut.moneyService.getMoney(c));
//    }

    @Test
    @PrepareForTest(DiscountsConfig.class)
    public void sellALot() {
        //given
        SellingService uut = new SellingService(persistenceLayer);
        Mockito.when(persistenceLayer.saveCustomer(Mockito.any())).thenReturn(Boolean.TRUE);
        Item i = new Item("i", new BigDecimal(3));
        Customer c = new Customer(1, "DasCustomer", "Kraków, Łojasiewicza");
        uut.moneyService.addMoney(c, new BigDecimal(990));
        PowerMockito.mockStatic(DiscountsConfig.class);
        PowerMockito.when(DiscountsConfig.isWeekendPromotion()).thenReturn(Boolean.FALSE);
        PowerMockito.when(DiscountsConfig.getDiscountForItem(i, c)).thenReturn(BigDecimal.ZERO);

        //when
        boolean sold = uut.sell(i, 10, c);

        //then
        Assert.assertFalse(sold);
        Assert.assertEquals(BigDecimal.valueOf(970),  uut.moneyService.getMoney(c));
    }
}


